Add palettes, glow, rules picker, population sparkline, and deeper zoom
Rules: famous-rules drop-down + per-neighbor B/S toggles replace the range sliders
Graphics: Viridis/Inferno/Plasma/Turbo palettes, softer cell edges, toggleable glow/bloom (G)
Gameplay: live population sparkline, Randomize button, Pulsar/Acorn presets, zoom-to-pattern Fit
Zoom: 1 GiB memory budget (was a 4M-cell cap); widen rule bitmasks to uint16_t
Docs: update README/NOTES to match; fix toolbar hover-label overflow

+++ src/gol.h
@@ -6,8 +6,8 @@
 #include <stddef.h>
 
 typedef struct {
-    uint8_t birth;      // bitmask: bit n set if birth on n neighbors
-    uint8_t survival;   // bitmask: bit n set if survive on n neighbors
+    uint16_t birth;     // bitmask: bit n set if birth on n neighbors (0..8)
+    uint16_t survival;  // bitmask: bit n set if survive on n neighbors (0..8)
 } GOLRules;
 
 // Cell bit packing: bit 0 = alive, bits 1..15 = age (0..32767)
